backend/requirements.txt
aiofiles==25.1.0
aiosmtplib==5.1.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asyncmy==0.2.11
asyncpg==0.31.0
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudscraper==1.2.71
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==9.0.0
Mako==1.3.10
MarkupSafe==3.0.3
msal==1.35.1
numpy==2.4.3
Office365-REST-Python-Client==2.6.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.1
passlib==1.7.4
pillow==12.1.1
playwright==1.58.0
plyer==2.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyee==13.0.1
PyJWT==2.12.1
PyMySQL==1.1.2
pyparsing==3.3.2
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.22
pytz==2026.1.post1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
selenium==4.41.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.6.3
uvicorn==0.42.0
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
zipp==3.23.0
